Separate out shared renderer functions
