+++ source/psx/renderer.c
@@ -248,11 +248,6 @@ void renderer_draw_mesh_shaded(const mesh_t* mesh, const transform_t* model_tran
 	PopMatrix();
 }
 
-void renderer_draw_mesh_shaded_offset(const mesh_t* mesh, const transform_t* model_transform, int tex_id_offset) {
-    tex_id_start = tex_id_offset;
-    renderer_draw_mesh_shaded(mesh, model_transform);
-}
-
 void renderer_draw_mesh_shaded_offset_local(const mesh_t* mesh, const transform_t* model_transform, int tex_id_offset) {
     tex_id_start = tex_id_offset;
     
@@ -299,14 +294,6 @@ void renderer_draw_mesh_shaded_offset_local(const mesh_t* mesh, const transform_
     }
 
     PopMatrix();
-}
-
-void renderer_draw_2d_quad_axis_aligned(vec2_t center, vec2_t size, vec2_t uv_tl, vec2_t uv_br, pixel32_t color, int depth, int texture_id, int is_page) {
-    const vec2_t tl = {center.x - size.x/2, center.y - size.y/2};
-    const vec2_t tr = {center.x + size.x/2, center.y - size.y/2};
-    const vec2_t bl = {center.x - size.x/2, center.y + size.y/2};
-    const vec2_t br = {center.x + size.x/2, center.y + size.y/2};
-    renderer_draw_2d_quad(tl, tr, bl, br, uv_tl, uv_br, color, depth, texture_id, is_page);
 }
 
 void renderer_draw_2d_quad(vec2_t tl, vec2_t tr, vec2_t bl, vec2_t br, vec2_t uv_tl, vec2_t uv_br, pixel32_t color, int depth, int texture_id, int is_page) {
@@ -480,33 +467,6 @@ void renderer_debug_draw_line(vec3_t v0, vec3_t v1, pixel32_t color, const trans
 
     PopMatrix();
 }
-
-void renderer_debug_draw_aabb(const aabb_t* box, pixel32_t color, const transform_t* model_transform) {
-    // Create 8 vertices
-    const vec3_t vertex000 = { box->min.x, box->min.y, box->min.z };
-    const vec3_t vertex001 = { box->min.x, box->min.y, box->max.z };
-    const vec3_t vertex010 = { box->min.x, box->max.y, box->min.z };
-    const vec3_t vertex011 = { box->min.x, box->max.y, box->max.z };
-    const vec3_t vertex100 = { box->max.x, box->min.y, box->min.z };
-    const vec3_t vertex101 = { box->max.x, box->min.y, box->max.z };
-    const vec3_t vertex110 = { box->max.x, box->max.y, box->min.z };
-    const vec3_t vertex111 = { box->max.x, box->max.y, box->max.z };
-
-    // Draw the lines
-    renderer_debug_draw_line(vertex000, vertex100, color, model_transform);
-    renderer_debug_draw_line(vertex100, vertex101, color, model_transform);
-    renderer_debug_draw_line(vertex101, vertex001, color, model_transform);
-    renderer_debug_draw_line(vertex001, vertex000, color, model_transform);
-    renderer_debug_draw_line(vertex010, vertex110, color, model_transform);
-    renderer_debug_draw_line(vertex110, vertex111, color, model_transform);
-    renderer_debug_draw_line(vertex111, vertex011, color, model_transform);
-    renderer_debug_draw_line(vertex011, vertex010, color, model_transform);
-    renderer_debug_draw_line(vertex000, vertex010, color, model_transform);
-    renderer_debug_draw_line(vertex100, vertex110, color, model_transform);
-    renderer_debug_draw_line(vertex101, vertex111, color, model_transform);
-    renderer_debug_draw_line(vertex001, vertex011, color, model_transform);
-}
-void renderer_debug_draw_sphere(sphere_t sphere) {}
 
 void renderer_upload_texture(const texture_cpu_t* texture, uint8_t index) {
     // Load texture pixels to VRAM - starting from 0,256, spanning 512x256 VRAM pixels, stored in 16x64 blocks (for 64x64 texture)
