+++ source/windows/renderer.c
@@ -619,43 +619,6 @@ void renderer_debug_draw_bvh_triangles(const level_collision_t* box, const pixel
     }
 }
 
-void renderer_debug_draw_aabb(const aabb_t* box, const pixel32_t color, const transform_t* model_transform) {
-    // Create 8 vertices
-    const vec3_t vertex000 = {box->min.x, box->min.y, box->min.z};
-    const vec3_t vertex001 = {box->min.x, box->min.y, box->max.z};
-    const vec3_t vertex010 = {box->min.x, box->max.y, box->min.z};
-    const vec3_t vertex011 = {box->min.x, box->max.y, box->max.z};
-    const vec3_t vertex100 = {box->max.x, box->min.y, box->min.z};
-    const vec3_t vertex101 = {box->max.x, box->min.y, box->max.z};
-    const vec3_t vertex110 = {box->max.x, box->max.y, box->min.z};
-    const vec3_t vertex111 = {box->max.x, box->max.y, box->max.z};
-
-    // Draw the lines
-    renderer_debug_draw_line(vertex000, vertex100, color, model_transform);
-    renderer_debug_draw_line(vertex100, vertex101, color, model_transform);
-    renderer_debug_draw_line(vertex101, vertex001, color, model_transform);
-    renderer_debug_draw_line(vertex001, vertex000, color, model_transform);
-    renderer_debug_draw_line(vertex010, vertex110, color, model_transform);
-    renderer_debug_draw_line(vertex110, vertex111, color, model_transform);
-    renderer_debug_draw_line(vertex111, vertex011, color, model_transform);
-    renderer_debug_draw_line(vertex011, vertex010, color, model_transform);
-    renderer_debug_draw_line(vertex000, vertex010, color, model_transform);
-    renderer_debug_draw_line(vertex100, vertex110, color, model_transform);
-    renderer_debug_draw_line(vertex101, vertex111, color, model_transform);
-    renderer_debug_draw_line(vertex001, vertex011, color, model_transform);
-}
-
-void renderer_debug_draw_sphere(const sphere_t sphere) {
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_from_int32s(sphere.radius, 0, 0)), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_from_int32s(-sphere.radius, 0, 0)), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_from_int32s(0, 0, sphere.radius)), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_from_int32s(0, 0, -sphere.radius)), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_mul(vec3_from_int32s(+sphere.radius, 0, +sphere.radius), vec3_from_scalar(2896))), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_mul(vec3_from_int32s(+sphere.radius, 0, -sphere.radius), vec3_from_scalar(2896))), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_mul(vec3_from_int32s(-sphere.radius, 0, +sphere.radius), vec3_from_scalar(2896))), white, &id_transform);
-    renderer_debug_draw_line(sphere.center, vec3_add(sphere.center, vec3_mul(vec3_from_int32s(-sphere.radius, 0, -sphere.radius), vec3_from_scalar(2896))), white, &id_transform);
-}
-
 void renderer_upload_texture(const texture_cpu_t *texture, const uint8_t index) {
 	// This is where all the pixels will be stored
 	pixel32_t *pixels = mem_alloc((size_t)texture->width * (size_t)texture->height * 4, MEM_CAT_TEXTURE);
@@ -739,15 +702,6 @@ int renderer_get_level_section_from_position(const model_t* model, vec3_t positi
         }
     }
     return n_sections; // -1 means no section
-}
-
-// todo
-void renderer_draw_2d_quad_axis_aligned(vec2_t center, vec2_t size, vec2_t uv_tl, vec2_t uv_br, pixel32_t color, int depth, int texture_id, int is_page) {
-	const vec2_t tl = { center.x - size.x / 2, center.y - size.y / 2 };
-	const vec2_t tr = { center.x + size.x / 2, center.y - size.y / 2 };
-	const vec2_t bl = { center.x - size.x / 2, center.y + size.y / 2 };
-	const vec2_t br = { center.x + size.x / 2, center.y + size.y / 2 };
-	renderer_draw_2d_quad(tl, tr, bl, br, uv_tl, uv_br, color, depth, texture_id, is_page);
 }
 
 void renderer_draw_2d_quad(vec2_t tl, vec2_t tr, vec2_t bl, vec2_t br, vec2_t uv_tl, vec2_t uv_br, pixel32_t color,
@@ -972,10 +926,6 @@ void renderer_upload_8bit_texture_page(const texture_cpu_t* texture, const uint8
 
 void renderer_set_video_mode(int is_pal) {}
 void renderer_cycle_res_x(void) {}
-void renderer_draw_mesh_shaded_offset(const mesh_t* mesh, const transform_t* model_transform, int tex_id_offset) {
-    tex_id_start = tex_id_offset;
-    renderer_draw_mesh_shaded(mesh, model_transform);
-}
 
 void renderer_draw_mesh_shaded_offset_local(const mesh_t* mesh, const transform_t* model_transform, int tex_id_offset) {
     tex_id_start = tex_id_offset;
